app/ccnFileProxy: In writeFile, fix timeout bug by changing to spawn a separate thread to write the output
writeFile was trying to send the data on the same thread for servicing
incoming interests and was getting blocked.  When it receives the
interest, ccnfileproxy gets the file and writes out all the segments
which are ultimately stored in the CCNFlowControl object.  When it tries
close the output, this eventually calls CCNFlowControl.waitForPutDrain
which expects all its output content to have been requested and
serviced.  But because this wait loop is on the same interest service
thread, no interests are serviced and the attempt to close times out
after 10 seconds.  This returns control to the interest service thread
and the content is sent.  The solution is to spawn a separate thread to
write the output segments and not block the incoming interest service
thread.

Refs #5